refactor: 유틸 함수 제거 후 fetch 함수 하나 생성 JUL-50
